Added more flexible commit split pattern
This splits commits on newline or start of the log, which must be
followed by a string of non-newline and non-start-token characters
followed by the start token.

This is more flexible because it allows for anything - not just graph
branch characters - before the start token and after the end token. This
allows for options like --stat and --left-right to work.