Allow specifying -raw-args multiple times
This allows for multiple scenarios.

The first is that the user adds default arguments to
g:flog_default_arguments but they still want to be able to pass in
additional args.

The second is that they previously set -raw-args on the command line,
but they want to now use -- to specify additional arguments without
going back to correct the command line.